common: move extent* files and os_badblocks_(count|get) to libpmem2
