+++ src/common/badblock_none.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2018, Intel Corporation */
+/* Copyright 2018-2020, Intel Corporation */
 
 /*
  * badblock_none.c - fake bad block API
@@ -20,33 +20,6 @@
  */
 int
 os_badblocks_check_file(const char *file)
-{
-	LOG(3, "file %s", file);
-
-	/* not supported */
-	errno = ENOTSUP;
-	return -1;
-}
-
-/*
- * os_badblocks_count -- returns number of bad blocks in the file
- *                       or -1 in case of an error
- */
-long
-os_badblocks_count(const char *file)
-{
-	LOG(3, "file %s", file);
-
-	/* not supported */
-	errno = ENOTSUP;
-	return -1;
-}
-
-/*
- * os_badblocks_get -- returns list of bad blocks in the file
- */
-int
-os_badblocks_get(const char *file, struct badblocks *bbs)
 {
 	LOG(3, "file %s", file);
