+++ src/libpmem2/badblocks_ndctl.c
@@ -17,6 +17,9 @@
 #include <ndctl/libndctl.h>
 #include <ndctl/libdaxctl.h>
 
+#include "libpmem2.h"
+#include "pmem2_utils.h"
+
 #include "file.h"
 #include "out.h"
 #include "os.h"
@@ -24,9 +27,7 @@
 #include "os_badblock.h"
 #include "badblock.h"
 #include "vec.h"
-
-#include "libpmem2.h"
-#include "pmem2_utils.h"
+#include "extent.h"
 
 #define FOREACH_BUS_REGION_NAMESPACE(ctx, bus, region, ndns)	\
 	ndctl_bus_foreach(ctx, bus)				\
@@ -708,4 +709,186 @@ badblocks_devdax_clear_badblocks_all(const char *path)
 	LOG(3, "path %s", path);
 
 	return badblocks_devdax_clear_badblocks(path, NULL);
+}
+
+/*
+ * badblocks_get -- returns 0 and bad blocks in the 'bbs' array
+ *                     (that has to be pre-allocated)
+ *                     or -1 in case of an error
+ */
+int
+badblocks_get(const char *file, struct badblocks *bbs)
+{
+	LOG(3, "file %s badblocks %p", file, bbs);
+
+	ASSERTne(bbs, NULL);
+
+	VEC(bbsvec, struct bad_block) bbv = VEC_INITIALIZER;
+	struct extents *exts = NULL;
+	long extents = 0;
+
+	unsigned long long bb_beg;
+	unsigned long long bb_end;
+	unsigned long long bb_len;
+	unsigned long long bb_off;
+	unsigned long long ext_beg;
+	unsigned long long ext_end;
+	unsigned long long not_block_aligned;
+
+	int bb_found = -1; /* -1 means an error */
+
+	memset(bbs, 0, sizeof(*bbs));
+
+	if (badblocks_files_namespace_badblocks(file, bbs)) {
+		LOG(1, "checking the file for bad blocks failed -- '%s'", file);
+		goto error_free_all;
+	}
+
+	if (bbs->bb_cnt == 0) {
+		bb_found = 0;
+		goto exit_free_all;
+	}
+
+	exts = Zalloc(sizeof(struct extents));
+	if (exts == NULL) {
+		ERR("!Zalloc");
+		goto error_free_all;
+	}
+
+	extents = os_extents_count(file, exts);
+	if (extents < 0) {
+		LOG(1, "counting file's extents failed -- '%s'", file);
+		goto error_free_all;
+	}
+
+	if (extents == 0) {
+		/* dax device has no extents */
+		bb_found = (int)bbs->bb_cnt;
+
+		for (unsigned b = 0; b < bbs->bb_cnt; b++) {
+			LOG(4, "bad block found: offset: %llu, length: %u",
+				bbs->bbv[b].offset,
+				bbs->bbv[b].length);
+		}
+
+		goto exit_free_all;
+	}
+
+	exts->extents = Zalloc(exts->extents_count * sizeof(struct extent));
+	if (exts->extents == NULL) {
+		ERR("!Zalloc");
+		goto error_free_all;
+	}
+
+	if (os_extents_get(file, exts)) {
+		LOG(1, "getting file's extents failed -- '%s'", file);
+		goto error_free_all;
+	}
+
+	bb_found = 0;
+
+	for (unsigned b = 0; b < bbs->bb_cnt; b++) {
+
+		bb_beg = bbs->bbv[b].offset;
+		bb_end = bb_beg + bbs->bbv[b].length - 1;
+
+		for (unsigned e = 0; e < exts->extents_count; e++) {
+
+			ext_beg = exts->extents[e].offset_physical;
+			ext_end = ext_beg + exts->extents[e].length - 1;
+
+			/* check if the bad block overlaps with file's extent */
+			if (bb_beg > ext_end || ext_beg > bb_end)
+				continue;
+
+			bb_found++;
+
+			bb_beg = (bb_beg > ext_beg) ? bb_beg : ext_beg;
+			bb_end = (bb_end < ext_end) ? bb_end : ext_end;
+			bb_len = bb_end - bb_beg + 1;
+			bb_off = bb_beg + exts->extents[e].offset_logical
+					- exts->extents[e].offset_physical;
+
+			LOG(10,
+				"bad block found: physical offset: %llu, length: %llu",
+				bb_beg, bb_len);
+
+			/* make sure offset is block-aligned */
+			not_block_aligned = bb_off & (exts->blksize - 1);
+			if (not_block_aligned) {
+				bb_off -= not_block_aligned;
+				bb_len += not_block_aligned;
+			}
+
+			/* make sure length is block-aligned */
+			bb_len = ALIGN_UP(bb_len, exts->blksize);
+
+			LOG(4,
+				"bad block found: logical offset: %llu, length: %llu",
+				bb_off, bb_len);
+
+			/*
+			 * Form a new bad block structure with offset and length
+			 * expressed in bytes and offset relative
+			 * to the beginning of the file.
+			 */
+			struct bad_block bb;
+			bb.offset = bb_off;
+			bb.length = (unsigned)(bb_len);
+			/* unknown healthy replica */
+			bb.nhealthy = NO_HEALTHY_REPLICA;
+
+			/* add the new bad block to the vector */
+			if (VEC_PUSH_BACK(&bbv, bb)) {
+				VEC_DELETE(&bbv);
+				bb_found = -1;
+				goto error_free_all;
+			}
+		}
+	}
+
+error_free_all:
+	Free(bbs->bbv);
+	bbs->bbv = NULL;
+	bbs->bb_cnt = 0;
+
+exit_free_all:
+	if (exts) {
+		Free(exts->extents);
+		Free(exts);
+	}
+
+	if (extents > 0 && bb_found > 0) {
+		bbs->bbv = VEC_ARR(&bbv);
+		bbs->bb_cnt = (unsigned)VEC_SIZE(&bbv);
+
+		LOG(10, "number of bad blocks detected: %u", bbs->bb_cnt);
+
+		/* sanity check */
+		ASSERTeq((unsigned)bb_found, bbs->bb_cnt);
+	}
+
+	return (bb_found >= 0) ? 0 : -1;
+}
+
+/*
+ * badblocks_count -- returns number of bad blocks in the file
+ *                       or -1 in case of an error
+ */
+long
+badblocks_count(const char *file)
+{
+	LOG(3, "file %s", file);
+
+	struct badblocks *bbs = badblocks_new();
+	if (bbs == NULL)
+		return -1;
+
+	int ret = badblocks_get(file, bbs);
+
+	long count = (ret == 0) ? (long)bbs->bb_cnt : -1;
+
+	badblocks_delete(bbs);
+
+	return count;
 }
