+++ src/libpmem2/extent_freebsd.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2018, Intel Corporation */
+/* Copyright 2018-2020, Intel Corporation */
 
 /*
  * extent_freebsd.c - implementation of the FreeBSD fs extent query API
