+++ src/libpmem2/extent_linux.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2018-2019, Intel Corporation */
+/* Copyright 2018-2020, Intel Corporation */
 
 /*
  * extent_linux.c - implementation of the linux fs extent query API
@@ -10,6 +10,9 @@
 #include <sys/ioctl.h>
 #include <linux/fs.h>
 #include <linux/fiemap.h>
+
+#include "libpmem2.h"
+#include "pmem2_utils.h"
 
 #include "file.h"
 #include "out.h"
@@ -35,13 +38,24 @@ os_extents_common(const char *path, struct extents *exts,
 		return -1;
 	}
 
-	enum file_type type = util_fd_get_type(fd);
-	if (type < 0)
-		goto error_close;
-
-	struct stat st;
-	if (fstat(fd, &st) < 0) {
+	os_stat_t st;
+	if (os_fstat(fd, &st) < 0) {
 		ERR("!fstat %d", fd);
+		goto error_close;
+	}
+
+	enum pmem2_file_type pmem2_type;
+
+	int ret = pmem2_get_type_from_stat(&st, &pmem2_type);
+	if (ret) {
+		errno = pmem2_err_to_errno(ret);
+		goto error_close;
+	}
+
+	/* directories do not have any extents */
+	if (pmem2_type == PMEM2_FTYPE_DIR) {
+		ERR(
+			"checking extents does not make sense in case of directories");
 		goto error_close;
 	}
 
@@ -51,10 +65,12 @@ os_extents_common(const char *path, struct extents *exts,
 	}
 
 	/* devdax does not have any extents */
-	if (type == TYPE_DEVDAX) {
+	if (pmem2_type == PMEM2_FTYPE_DEVDAX) {
 		close(fd);
 		return 0;
 	}
+
+	ASSERTeq(pmem2_type, PMEM2_FTYPE_REG);
 
 	struct fiemap *fmap = Zalloc(sizeof(struct fiemap));
 	if (fmap == NULL) {
