+++ src/test/util_extent/Makefile
@@ -1,10 +1,15 @@
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright 2018, Intel Corporation
+# Copyright 2018-2020, Intel Corporation
 
 #
 # src/test/util_extent/Makefile -- build util_extent test
 #
 #
+
+TOP = ../../..
+
+INCS += -I$(TOP)/src/libpmem2
+
 TARGET = util_extent
 OBJS = util_extent.o
